Just import shared `PropType`s as an object for now
