+++ ui/js/pods-dfv/_src/core/react/pods-dfv-prop-types.js
@@ -1,0 +1,6 @@
+import PropTypes from 'prop-types';
+
+export const PodsDFVPropTypes = {
+	fieldType: PropTypes.string,
+	fieldItemData: PropTypes.array
+};
