+++ ui/js/pods-dfv/_src/text/text.js
@@ -1,13 +1,25 @@
 import React from 'react';
-import { PodsDFVReactComponent } from 'pods-dfv/_src/core/react/pods-dfv-react-component';
+import PropTypes from 'prop-types';
+import { PodsDFVPropTypes } from 'pods-dfv/_src/core/react/pods-dfv-prop-types';
 
 
-export class PodsDFVText extends PodsDFVReactComponent {
+export class PodsDFVText extends React.Component {
 	constructor ( props ) {
 		super( props );
 		this.state = {
 			value: this.props.fieldItemData[ 0 ]
 		};
+	}
+
+	static get propTypes () {
+		return {
+			...PodsDFVPropTypes,
+			fieldConfig: {
+				text_max_length: PropTypes.string,
+				text_placeholder: PropTypes.string,
+				readonly: PropTypes.bool
+			}
+		}
 	}
 
 	onChanged ( event ) {
